+++ docs/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 
 * Addition of the :mod:`~nenupy.observation.obs_config` module to estimate data volumes of NenuFAR observations. This can be done for all the receivers and configuration modes. Observation configurations can either be set individualy for each receiver or given as parset files or list of parset files [`#43 <https://github.com/AlanLoh/nenupy/issues/43>`_].
 * Module :mod:`~nenupy.schedule` to automatically schedule observation bookings via deterministic or genetic algorithm.
+* Method :func:`~nenupy.instru.lnaGain` to get the NenuFAR Low Noise Amplifier gains interpolated at given frequencies.
 
 
 1.1.0 (2020-11-18)